backend/requirements.txt
Flask
transformers==4.33.2
phonemizer
librosa
soundfile
pydub
flask-cors
torch
numpy<2
marshmallow
mysql-connector-python
